requirements.txt
fastapi==0.111.2
uvicorn==0.24.0
pandas==2.1.1
scikit-learn==1.3.3
pyyaml==6.0
openpyxl==3.1.3
